restructure async module as directory with `async-io` and `tokio` submodules
